Move HipFileDataOps to its own file.
